[#156] Refactoring, minor fixes
