--- etl/analyze_code.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-pycodestyle etl
-pycodestyle tests
-pydocstyle etl
-pylint etl
