+++ legion_airflow/MANIFEST.in
@@ -1,2 +1,2 @@
-graft etl/slack
+graft legion_airflow/slack
 recursive-include requirements *.txt
