+++ legion_airflow/analyze_code.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+pycodestyle legion_airflow
+pycodestyle tests
+pydocstyle legion_airflow
+pylint legion_airflow
