+++ legion_airflow/setup.cfg
